test: reorganization of the 'info' unit test
