+++ tests/unit/CMakeLists.txt
@@ -8,7 +8,7 @@ add_subdirectory(conn_req)
 add_subdirectory(ep-test)
 add_subdirectory(error)
 add_subdirectory(flush-test)
-add_subdirectory(info-test)
+add_subdirectory(info)
 add_subdirectory(log)
 add_subdirectory(mr)
 add_subdirectory(peer)
