+++ tests/unit/common/mocks-rdma_cm.c
@@ -14,6 +14,7 @@
 #include "conn_req.h"
 #include "mocks-ibverbs.h"
 #include "mocks-rdma_cm.h"
+#include "test-common.h"
 
 struct rdma_event_channel Evch; /* mock event channel */
 struct rdma_cm_id Cm_id;	/* mock CM ID */
@@ -26,6 +27,8 @@ int Rdma_migrate_id_counter = 0;
 
 /* mock control entity */
 int Mock_ctrl_defer_destruction = MOCK_CTRL_NO_DEFER;
+
+const struct rdma_cm_id Cmid_zero = {0};
 
 /*
  * rdma_create_qp -- rdma_create_qp() mock
@@ -287,5 +290,79 @@ rdma_get_cm_event(struct rdma_event_channel *channel,
 	}
 
 	*event_ptr = event;
+	return 0;
+}
+
+/*
+ * rdma_getaddrinfo -- rdma_getaddrinfo() mock
+ */
+int
+rdma_getaddrinfo(const char *node, const char *service,
+		const struct rdma_addrinfo *hints, struct rdma_addrinfo **res)
+{
+	struct rdma_addrinfo_args *args =
+				mock_type(struct rdma_addrinfo_args *);
+	if (args->validate_params == MOCK_VALIDATE) {
+		assert_string_equal(node, MOCK_IP_ADDRESS);
+		assert_string_equal(service, MOCK_SERVICE);
+		check_expected(hints->ai_flags);
+	}
+
+	*res = args->res;
+
+	if (*res != NULL)
+		return 0;
+
+	errno = mock_type(int);
+
+	return -1;
+}
+
+/*
+ * rdma_freeaddrinfo -- rdma_freeaddrinfo() mock
+ */
+void
+rdma_freeaddrinfo(struct rdma_addrinfo *res)
+{
+	struct rdma_addrinfo_args *args =
+				mock_type(struct rdma_addrinfo_args *);
+	if (args->validate_params == MOCK_VALIDATE)
+		assert_ptr_equal(res, args->res);
+}
+
+/*
+ * rdma_resolve_addr -- rdma_resolve_addr() mock
+ * Note: CM ID is not modified.
+ */
+int
+rdma_resolve_addr(struct rdma_cm_id *id, struct sockaddr *src_addr,
+		struct sockaddr *dst_addr, int timeout_ms)
+{
+	check_expected_ptr(id);
+	check_expected_ptr(src_addr);
+	check_expected_ptr(dst_addr);
+	check_expected(timeout_ms);
+
+	errno = mock_type(int);
+	if (errno)
+		return -1;
+
+	return 0;
+}
+
+/*
+ * rdma_bind_addr -- rdma_bind_addr() mock
+ * Note: CM ID is not modified.
+ */
+int
+rdma_bind_addr(struct rdma_cm_id *id, struct sockaddr *addr)
+{
+	check_expected_ptr(id);
+	check_expected_ptr(addr);
+
+	errno = mock_type(int);
+	if (errno)
+		return -1;
+
 	return 0;
 }
