+++ tests/unit/common/mocks-rdma_cm.h
@@ -40,4 +40,11 @@ extern int Rdma_migrate_id_counter;
  */
 extern int Mock_ctrl_defer_destruction;
 
+struct rdma_addrinfo_args {
+	int validate_params;
+	struct rdma_addrinfo *res;
+};
+
+extern const struct rdma_cm_id Cmid_zero;
+
 #endif /* MOCKS_RDMA_CM_H */
